+++ common/pf_log.h
@@ -18,7 +18,7 @@
  * (eay@cryptsoft.com).  This product includes software written by Tim
  * Hudson (tjh@cryptsoft.com).
  *
- * $Id$
+ * $Id: pf_log.h 1138 2008-06-08 12:22:09Z romain $
  */
 
 #ifndef PF_LOG_H
@@ -74,7 +74,7 @@ public:
 	class flux
 	{
 		std::string str;
-		uint32_t flag;
+		size_t flag;
 
 		public:
 			flux(size_t i)
